adjust: separate Search results component (previously BookList) for reuse
turned BookList in a reusable component, that is now used in the SearchResults component
