+++ src/app/(public)/list/[listId]/page.tsx
@@ -1,0 +1,7 @@
+export default function Page({ params }: { params: { listId: string } }) {
+    return (
+        <>
+        <div>{params.listId}</div>
+        </>
+    )
+};
